+++ packages/components/src/slot-fill/fill.js
@@ -6,19 +6,17 @@ import { isFunction } from 'lodash';
 /**
  * WordPress dependencies
  */
-import { createPortal, useLayoutEffect, useRef, useState } from '@wordpress/element';
+import { createPortal, useLayoutEffect, useRef } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
-import { Consumer } from './context';
+import { Consumer, useSlot } from './context';
 
 let occurrences = 0;
 
-function FillComponent( { name, getSlot, children, registerFill, unregisterFill } ) {
-	// Random state used to rerender the component if needed, ideally we don't need this
-	const [ , updateRerenderState ] = useState( {} );
-	const rerender = () => updateRerenderState( {} );
+function FillComponent( { name, children, registerFill, unregisterFill } ) {
+	const slot = useSlot( name );
 
 	const ref = useRef( {
 		name,
@@ -30,14 +28,12 @@ function FillComponent( { name, getSlot, children, registerFill, unregisterFill 
 	}
 
 	useLayoutEffect( () => {
-		ref.current.forceUpdate = rerender;
 		registerFill( name, ref.current );
 		return () => unregisterFill( name, ref.current );
 	}, [] );
 
 	useLayoutEffect( () => {
 		ref.current.children = children;
-		const slot = getSlot( name );
 		if ( slot && ! slot.props.bubblesVirtually ) {
 			slot.forceUpdate();
 		}
@@ -53,8 +49,6 @@ function FillComponent( { name, getSlot, children, registerFill, unregisterFill 
 		registerFill( name, ref.current );
 	}, [ name ] );
 
-	const slot = getSlot( name );
-
 	if ( ! slot || ! slot.node || ! slot.props.bubblesVirtually ) {
 		return null;
 	}
@@ -69,10 +63,9 @@ function FillComponent( { name, getSlot, children, registerFill, unregisterFill 
 
 const Fill = ( props ) => (
 	<Consumer>
-		{ ( { getSlot, registerFill, unregisterFill } ) => (
+		{ ( { registerFill, unregisterFill } ) => (
 			<FillComponent
 				{ ...props }
-				getSlot={ getSlot }
 				registerFill={ registerFill }
 				unregisterFill={ unregisterFill }
 			/>