+++ packages/core-data/src/queried-data/get-query-parts.js
@@ -11,7 +11,7 @@ import { withWeakMapCache } from '../utils';
 /**
  * An object of properties describing a specific query.
  *
- * @typedef {WPQueriedDataQueryParts}
+ * @typedef {Object} WPQueriedDataQueryParts
  *
  * @property {number} page      The query page (1-based index, default 1).
  * @property {number} perPage   Items per page for query (default 10).