+++ packages/eslint-plugin/configs/recommended.js
@@ -2,9 +2,10 @@ module.exports = {
 	parser: 'babel-eslint',
 	extends: [
 		require.resolve( './jsx-a11y.js' ),
-		require.resolve( './react.js' ),
 		require.resolve( './custom.js' ),
+		require.resolve( './react.js' ),
 		require.resolve( './esnext.js' ),
+		require.resolve( './jsdoc.js' ),
 	],
 	env: {
 		node: true,