+++ packages/notices/src/store/selectors.js
@@ -15,7 +15,7 @@ import { DEFAULT_CONTEXT } from './constants';
 const DEFAULT_NOTICES = [];
 
 /**
- * Notice object.
+ * @typedef {Object} WPNotice Notice object.
  *
  * @property {string}  id               Unique identifier of notice.
  * @property {string}  status           Status of notice, one of `success`,
@@ -31,11 +31,10 @@ const DEFAULT_NOTICES = [];
  *                                      user. Defaults to `true`.
  * @property {WPNoticeAction[]} actions User actions to present with notice.
  *
- * @typedef {WPNotice}
  */
 
 /**
- * Object describing a user action option associated with a notice.
+ * @typedef {Object} WPNoticeAction Object describing a user action option associated with a notice.
  *
  * @property {string}    label    Message to use as action label.
  * @property {?string}   url      Optional URL of resource if action incurs
@@ -43,7 +42,6 @@ const DEFAULT_NOTICES = [];
  * @property {?Function} callback Optional function to invoke when action is
  *                                triggered by user.
  *
- * @typedef {WPNoticeAction}
  */
 
 /**