+++ packages/babel-preset-default/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 4.4.0 (2019-08-05)
+
+### Bug Fixes
+
+- Configure Babel to target your current version of Node as described in [Jest docs](https://jestjs.io/docs/en/getting-started#using-babel).
+- Added missing [core-js](https://www.npmjs.com/package/core-js) dependency ([#16259](https://github.com/WordPress/gutenberg/pull/16259)).
+
 ## 4.2.0 (2019-05-21)
 
 ### New Features